requirements.txt
fastapi==0.110.0
uvicorn[standard]
pydantic>=2.0
python-dotenv
httpx
tortoise-orm
jinja2